Move AccountTable into separate component
